+++ R/dynR-package.R
@@ -1,0 +1,14 @@
+#' dynR: Dynamic Connectivity Analysis for Neurophysiological Timeseries
+#'
+#' An R port of the Python dynfc library for computing dynamic functional
+#' connectivity representations from multivariate neurophysiological
+#' timeseries, including BOLD fMRI, EEG, LFP, and related signals.
+#'
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @useDynLib dynR, .registration = TRUE
+#' @importFrom Rcpp sourceCpp
+## usethis namespace: end
+NULL